select dropdown js: select blank values
This updates selects to fix a bug with selecting blank values.
When a select dropdown is initialized, it would set its prev
value to a blank string (`''`). When selecting an option, if its
value was also a blank string, then it would fail to "select" the
option. This is because it checks the prev value and new value
and only selects an option if they don't match. In the above
scenario, they do match, so it wouldn't select the blank option.
To fix this, the prev value is defaulted to `null`, which allows
the comparison of the prev and new value to work as expected.