select btn group: append caret in the select btn, not the wrapper
When using the select inline with another `romo-btn-group` you specify
an option that makes the wrapper a `romo-btn-group`.  But you can't
have wrapper children that aren't `.romo-btn` b/c the border handling
messes up.

I originally tried to fix this by altering the border handling
selector but this has proven to only break normal `romo-btn-group`
handling.  I've reverted that change here.

Instead, I now add the carent icon to the select romo-btn instead
of the wrapper.  This has the same effect as before while not
breaking the case where the wrapper is a `romo-btn-group`.