form js: add `submitCompleted` event
This event will fire when the submit is finished, whether it was
successful or errored.  This is to make the API and feature set
more rich and so you can trigger behavior once the submit completes.

This also cleans up the indicator handling.  First it moves to triggering
the indicator before firing the before submit event.  This ensures
the very first and very last things done in the submit cylcle it
to start/stop the indicator.  This also updates the indicator logic
to run even if the element is disabled.  This is necessary as users
may wish to disable the indicator but still stop it later on.