dropdown js: set offset if setting position fixed
This fixes a bug in Chrome where it doesn't properly honor the offset
it the dropdown element is dynamically changed to position: fixed.
This comes up when using dropdowns (for example select dropdowns)
in modals (ie model forms).  The initial opening of the dropdown
sets position fixed and calculates the offset.

It seems the calculating the offset the first time "resets" how
chrome determines the offset (ie honor postion fixed) b/c subsequent
offset calcs are done correctly.

This is somewhat a "hack" in that this behavior is not exhibited
in Firefox.  It is less of a hack though because we only apply it
when setting position fixed dropdowns (ie only when it is needed).