dropdown js: rework max height detection; make it work for position top
This reworks the dropdown js to clean up how it detects the max
height available for the popup.  It also updates how this max height
applies to positioning the popup so that this calculation is honored
and works if positioning the popup on the top (previously positioning
top wouldn't benefit from max height detection).

This is prep for dropdowns switching their position if there isn't
enough height available for their popup.