js: kepress component
This component binds to an element and detects and publishes trigger
events on keypress events.  It is auto-invoked like other romo
components, publishes events and generally behaves like other romo
components.

This is a utility component.  It is designed to be used as a part
of bigger js components or behavior.  For example:

* submit-as-you-type form input
* pickers that search as you type
* keyboard shortcut handlers
* etc