form js: detect keypress submit elems and submit on keypress trigger
This updates forms to detect any elements that have been marked as
keypress submit elements.  It binds the keypress trigger event to
submitting the form.

Note: this adds a minimal delay timeout on the event binding.  This
avoids submitting the form on *every* keypress.  Instead this behaves
more like "submit the form when there is any pause in typing".