update base init helpers to manually init every elem in the given collection
When performing the init operation on multi-elem collections, only
one elem in the collection would actually get the auto-init behavior
triggered.  This updates to manually do both the operation and the
auto-init trigger on all given elements manually to ensure it happens
consistently to each.

As a side-effect, these methods no longer return the elements they
operate on.  They now return nothing.