css: add 'base' and 'alt' text/bg colors
This is in addition to the 'muted', 'warning', etc colors.  This
allows you to get to back to the base/alt colors when some parent
elem has some custom color set.  This is also useful to get back
to the base/alt on hover when non-hover has a custom color.