indicator js: capture css/html contents on start
This as opposed to on init.  Sometimes the full css width/height
etc aren't set at init-time (due to markup being lazy loaded, etc).
This allows the dimensions and markup to change and the captured
css/html to handle this appropriately.

This switches to conditionally setting the html on stop.  This case
has to be handled now that the html isn't being captured on init.

Finally, this switches to capturing the style attr it overrides
and setting it on stop.  This ensures it safely applies the css
necessary while the indicator is running but resets the state and
cleans things up appropriately.