dropdown: turn off body event handling if modal is opened
This updates the dropdown component to turn off its body event
handling while a modal (that was opened from the dropdown) is open.
The effect here is that any body event handling will be limited to
the modal while it is open and the dropdown will ignore body event
handling until the modal is closed.

This only applies to modals opened from dropdowns.  Modals opened
outside of the dropdown behave as before.

Note: this also switches dropdowns/tooltips to close on modal *moves*.
The old logic closed all popups as soon as you mouse down.  This is
technically not necessary as problems don't come up until you
start moving the modal (dragging it).  The changes to wait as long
as possible before closing.  This was noticed while messing around
with the above changes.