romo-hover style class concept; applied to display style classes
This introduces a new style class, `romo-hover` which is designed
to enable child hover style classes when the parent elem is hovered.
This also applies this concept to the display helper style classes.

The practical application here is to conditionally show child elems
when a parent is hovered - all using style classes instead of relying
on custom CSS.  To accomplish this example, you would add `romo-hover`
to the parent elem and then add `romo-display-none` and
`romo-display-{block|whatever}-hover` to the child elems that you
only want displayed on hover.  This would make the element hidden
by default but display {whatever} when the parent is hovered.