js: track parent-child elem relationships; use to remove appropriately
The problem we are tying to solve here is that some js components
that bind to DOM elems need to create additional DOM elems that
are not hierarchial children but are in their nature/design.  These
generated DOM elems are useless without the source elem.  However,
if the source elem was removed, these DOM elems would linger b/c
they existed in a separate elem tree.

This adds a parent/child elem tracking class that monitors DOM
removals at the body level.  If it detects a node removal that
contains parent elems, it goes and looks up the child elems for
each parent and removes them.  This keeps the DOM clean and avoids
weird edge-case behavior of orphaned DOM elements lingering when
they shouldn't.

This has been added to the tooltip, modal, dropdown, select and
datepicker components.  The tooltips, modals and dropdowns all
build popup elems that are appended to the body.  The selects appends
a wrapper after itself while the datepicker appends a wrapper and
then nests itself in said wrapper.  These wrappers/popups should
be removed if/when their component elems are removed.

The tracking class is a collection with `add` and `remove` methods.
Components use the `add` method to associate child elems with parent
elems.  The `remove` method is used by the DOM observer to remove
all child elems for parent elems in a node's tree.  Internally, the
tracking class generates elem ids and sets them for usage as a data
attr on the parent elems.