hotfix: add all parent-child elems on a delayed timeout
This is b/c components may be generated by other components and
appended around in weird ways.  Appends generate removals and re-adds
so we want to delay adding the child elem until after the removal
has been processed.  Specifically this was happening with selects
generating and appending dropdown elems.  The dropdown popup was
getting removed b/c it was being appended to the select wrapper.

This also switches to using a timeout of 1 ms.  Any value will do
(you just need a break so that the event handler code can run
before the child attrs are added).  It is preferrable to use the
shortest possible value to limit any timing weirdness.

/cc @jcredding