hotfix: modals/dropdowns unbind their underlying invokes (like tooltips)
In the case of modals/dropdowns, they manually handle invoking their
invokes.  This explicitly unbinds the underlying invoke so that it
can only be triggered programmatically.

Note: this is mostly precautionary.  In practice, the modal/dropdown
usage didn't cause double-invokes but we want to ensure that is
always the case.

/cc @jcredding