hotfix: don't propagate invoke events for composed invokes
If a component composes an invoke (ie modals, dropdowns, tooltips,
inlines, etc), it needs to prevent the invoke events from propagating
up.  If not, this prevents components that compose invokes from
being nested.  Without this both the inner and the outer components
will catch the event when it is only intended that the inner component
does.

The specific case I ran into was I wanted to put a confirmation
modal invoked from a form that was inside an inline (an inline
editable form).

/cc @jcredding