hotfix: modals/dropdowns: add `data-romo-modal-content-elem` attr
This allows you to specify an elem to use for the html contents of
the modal/dropdown popup.  On popup open, if an elem is specified
and exists, its contents will be copied into the popup body instead
of fetching the content using the invoke.

Use this when the content of the modal is static in place of specifying
hrefs or invoke attrs to fetch remote content.

/cc @jcredding