inline js: cleanup invoke handling; always show before init html
This does a couple of cleanups to how inlines handle interacting
with invoke components.  This all surfaced in light of changes made
in #54.  That PR inadvertently switched invokes to trigger the load
success event before triggering the general invoke event.  This
shouldn't have broken anything, but it had side-effects on inline
components.  Inlines used the general invoke event to trigger showing
the inline elem and used the load success event to trigger init'ing
the loaded html.  Inlines were assuming the general invoke would
always come in before the load success event (as you don't want to
init html that is hidden).  So when I changed the event order in
it (which isn't intended as components may use whether they are
shown or not to influence behavior based on init).

This really exposed some unintended coupling between inlines and
invokes.  If the inlines are picky about the state of the inline
when load success comes it, they should really handle it themselves.

This does a couple of changes to fix things:

* remove the `doInvoke` method on the inline.  A better name for this
  method is "show".  This updates the method to be called `doShow`
  and switches to triggering `inline:show` events.  This also helps
  with name confusing with the invoke events.  Also this more accurately
  describes what the method does (what does "invoking" an inline do?).
* update the load success/error logic to always show the inline before
  initializing the newly loaded html content.  The inline expects the
  inline contents to always be shown when initializing them.