hotfix: fix for indicator wiping width/height inline styles on init
This is somewhat convoluted.  Indicators, on init, bind a page show
event to stop themselves.  This handles a case where an indicator
was left spinning when a new page was loaded.  If you hit back the
indicator would still be spinning (b/c that is the state you left
the indicator in).

This behavior assumes that all the state is in place to stop the
indicator on page show.  This was not the case.  The start logic
captures some some state before mutating the UI to bring up the
spinner.  The stop logic then reads that state and returns the UI
to how it was before the indicator started.  This bug only affected
inline width, height or position styles. On start, custom versions
of these styles are needed so we save off any inline styles and set
the custom values.  On stop we clear the custom values and return
the inline styles.  If you stop before saving off any inline styles
the width, height and position will be cleared out on stop but won't
be restored as designed.

This simply saves off any inline styles on init so that stop is safe
to be called before start has been called.

/cc @jcredding