hotfix: add missing hover styles for striped and striped-alt tables
This was accidentally missed when originally implementing the table
and grid table styles.  This reason it was missed is that it just
happens to work with the "non alt" striped tables.  In the case of
the alt striped tables, their styles end up being more specific than
the hover styles and end up taking precedence where this doesn't
apply to the non-alt striped tables.

I'm not sure I ever explicitly used an alt striped table with hovering.
I noticed this b/c I finally used one in an app.