selects/datepickers: better handle dynamic carets/indicators
We allow you to set custom caret/indicator style classes (which in
turn have custom widths) but we were hard-coding `22px` of padding
to make room for the caret/indicator UI.  This is incosistent and
brittle as it is too easy to "cut off" the UI b/c the custom class
specified UI that is wider than 22px.

This adds additional API to both components that allows specifying
a width in pixels in additon to the style class.  Romo now calculates
the necessary padding to show the UI based on this custom value
(accounting for padding on the left/right sides).

Note: the "1px optical illusion I reference in the select code is
due to how Romo renders selects.  The select UI has a border that
is not at the true right edge of the element - there is an additional
1px "shadow".  The shadow is the actual edge of the element.  However,
our eyes see the border as the edge of the element so we need to
account for this in the padding/positioning calculations.

Note: this also moves defaulting the select caret values to the
select js file.  It was previously in the select dropdown file by
mistake.