(hotfix) formally bring in dassets, dassets-sass as dependencies
So Romo has built in Dassets integration.  I originally added this
to satisfy one of my project's needs quickly.  It was kinda
thrown in there with TODO comments to figure out how to better
handler integrating these gems once I had time to play around
with them together.  Well, there were some hidden problems
with this.

First off, I was requiring dassets and dassets-sass but never
specifying them as a formal dependencies in the gemspec.  They
were brought in formally in the Gemfile to build the gh-pages docs
but they weren't in the gemspec.  My project also hid this as it
brought in these gems in its Gemfile too.

Second, I required and  configured dassets-erb when it wasn't
needed.  I think I added this thinking I *would* need it later but
that hasn't happened yet.  If I need it in the future, I can
configure and add it back in then.

I noticed these issues when bringing in Romo on a greenfield
website project.