datepicker: honor timezone when determining today's date
This updates how today's date is calculated to properly honor
time zones.  Datepicker uses UTC dates for comparisons.  Before
we were assuming `new Date()` was a UTC date for today when in
fact it was a time zoned date.  This uses the timezoned date info
to build a UTC date for the date as perceived in the current time
zone.  Then this new UTC date is used for the "today" comparisons.