use `getComputedStyle` to lookup/parse elem z-index values
This is used to set appropriate cumulative z-index values for
modals/dropdowns/etc.  Chrome's latest update broke our old way
of getting an elem's z-index.  This switches to the more robust
and now fully supported `getComputedStyle` which works in all
browsers.

Note: I had to switch to ignoring `0` index values as this is the
default value of `getComputedStyle`.  Since we add zero on in the
end, the default effect is the same.