add a `toArray` helper global function
This commonizes a couple places where we've manually converted a
jQuery-style element collection to a vanilla js array of nodes.
I chose to implement this function in pure js instead of relying
on the jQuery-specific `toArray` method on collections.  I am
trying to move to using vanilla js whenever possible as we are
targeting modern browsers and they are getting pretty good these
days.

This will also be needed in a coming effort to implement a method
to do word-boundary filtering on collections of elements.