refine select next/prev option logic
This is prep for adding a filter for select options.  When testing
the filter in a grouped select, if you filtered such that some of
the groups were empty, it broke cycling to the prev or next opt.

This reworks the logic to both be more simple and easy to read,
but also work with empty nested option lists.  This new
implementation includes adding `selectPrev` and `selectNext`
Romo utility helper methods.  These behave like `$().prev` and
`$().next` but take a selector and only consider elems that
match the selector.  These are used to cycle to the next/prev
opts in the current list and if there is no opt in the current
list, then cycle to the prev/next list.

Note, this also includes a few minor handling tweaks that are
somewhat related:

* this disables the `tab` key while the popup is open.  this
  matches how native selects seem to work so I wanted to emulate
  it in Romo's select
* opens the popup on the 'Up' key as well as the 'Down' key.
  again, this is how the native selects seem to work so I want
  Romo's to emulate this.