(hotfix) update the gh-pages deas server dependencies
This gets the Deas/Dassets handling up to the latest conventions.
This is needed to rebuild the docs pages.  It had been a while
since this Deas app had been updated so things were out of date
a bit.  This gets the pages loading and gets on to the latest
dependency versions.